AI_Bot/requirements.txt
aiogram
openai
gspread
google-auth
python-dotenv